# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cours Online API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/google:
    post:
      operationId: googleLogin
      tags: [auth]
      summary: Sign in or sign up with Google
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleLoginInput"
      responses:
        "200":
          description: Google login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid Google credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/admin-login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Admin login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: User logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/me/phone:
    patch:
      operationId: updateMyPhone
      tags: [auth]
      summary: Set or update current user's WhatsApp phone number
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePhoneInput"
      responses:
        "200":
          description: Phone updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterInput:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        phone:
          type: string
          minLength: 9
      required:
        - username
        - email
        - password
        - phone
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    GoogleLoginInput:
      type: object
      properties:
        credential:
          type: string
      required:
        - credential
    UpdatePhoneInput:
      type: object
      properties:
        phone:
          type: string
          minLength: 9
      required:
        - phone
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    AdminAuthResponse:
      type: object
      properties:
        token:
          type: string
        admin:
          type: object
          properties:
            id:
              type: integer
            username:
              type: string
          required:
            - id
            - username
      required:
        - token
        - admin
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        accountType:
          type: string
          enum: [vip, normal]
        subscriptionType:
          type: string
          enum: [demo, monthly, annual, lifetime]
        subscriptionExpiresAt:
          type: string
          format: date-time
          nullable: true
        subscriptionIsExpired:
          type: boolean
        isActive:
          type: boolean
        phone:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
        communityRole:
          type: string
          enum: [admin, formateur, student]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - accountType
        - subscriptionType
        - communityRole
        - isActive
        - createdAt
    ChangePasswordInput:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6
      required:
        - currentPassword
        - newPassword
...