# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cours Online API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileAccountType } from "./userProfileAccountType";
import type { UserProfileCommunityRole } from "./userProfileCommunityRole";
import type { UserProfileSubscriptionType } from "./userProfileSubscriptionType";

export interface UserProfile {
  id: number;
  username: string;
  email: string;
  accountType: UserProfileAccountType;
  subscriptionType: UserProfileSubscriptionType;
  subscriptionExpiresAt?: Date | null;
  subscriptionIsExpired?: boolean;
  isActive: boolean;
  phone?: string | null;
  profileImageUrl?: string | null;
  communityRole: UserProfileCommunityRole;
  createdAt: Date;
}